+++ resources/OLEDB-Runner/OLEDB-Runner/obj/x64/Release/net6.0/OLEDB-Runner.AssemblyInfo.cs
@@ -12,7 +12,7 @@ using System;
 using System.Reflection;
 
 [assembly: System.Reflection.AssemblyCompanyAttribute("OLEDB-Runner")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
+[assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
 [assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+a5f893ff5f8e4d8ed28478ecfcd48ea2e9e8b313")]
 [assembly: System.Reflection.AssemblyProductAttribute("OLEDB-Runner")]